server/pyproject.toml
[project]
name = "reviv-pics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "boto3>=1.42.14",
    "celery>=5.6.0",
    "cloudinary>=1.44.1",
    "cryptography>=46.0.3",
    "dj-rest-auth[with-social]>=1.1.2",
    "django>=6.0",
    "django-allauth>=65.13.1",
    "django-cors-headers>=4.9.0",
    "django-ratelimit>=4.1.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.27.0",
    "fido2>=2.0.0",
    "gunicorn>=23.0.0",
    "jwt>=1.4.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "stripe>=14.1.0",
]

[tool.pyright]
reportMissingImports = false
reportMissingModuleSource = false
